+++ CHANGELOG.md
@@ -14,7 +14,8 @@ To publish a new release run `scripts/release` from the project directory.
 ### Fixed
 - Do not fail with broken pipe error when rending large images (see [GH-134] by
   [@fspillner]).
-  
+
+[GH-134]: https://github.com/lunaryorn/mdcat/issues/134  
 [GH-138]: https://github.com/lunaryorn/mdcat/issues/138
 
 ## [0.16.1] – 2020-05-15